Added Time to data
Added many changes to the save and load json function that allow for much easier control of the data. It's now possible to set a date of the Balance and it's possible to edit that date or add new or older values. The time history is now shown in the Cumulative budget plot
The bankBalance, when loaded, is updated starting from the earliest found date and is updated ascendingly up to a specified date. This allows easy transition through time.
The save function creates a new history item if a new date is saved, and if an existing date is modified, then the save function modifies it accordingly.
The save and load function also cleverly use the id number of a field to update it or maintain it over time